Make incremental sessions identity no longer depend on the crate names provided by source code
